+++ app/views/Precios/index.blade.php
@@ -2,9 +2,39 @@
 
 @section('main')
 
-<h1>All Precios</h1>
+<div class="page-header clearfix">
+    <h2 class="pull-left"><span class="glyphicon glyphicon-cog"></span> Precios </h2>
+    <div class="pull-right">
+        <a href="{{{ URL::route('Configuracion') }}}" class="btn btn-sm btn-success"><span class="glyphicon glyphicon-arrow-left"></span> Regresar</a>
+    </div>
+</div>
 
-<p>{{ link_to_route('Precios.create', 'Add new Precio') }}</p>
+<div class="btn-agregar">
+	<a type="button" href="{{ URL::route('Precios.create') }}" class="btn btn-primary">
+	  <span class="glyphicon glyphicon-file"></span> Nuevo Precio
+	</a>
+</div>
+
+@if ($errors->any())
+  <div class="alert alert-danger fade in">
+    <button type="button" class="close" data-dismiss="alert" aria-hidden="true">×</button>
+    @if($errors->count() > 1)
+      <h4>Oh no! Se encontraron errores!</h4>
+    @else
+      <h4>Oh no! Se encontró un error!</h4>
+    @endif
+    <ul>
+      {{ implode('', $errors->all('<li class="error">:message</li>')) }}
+    </ul>  
+  </div>
+@else
+	@if (Session::has('message'))
+		<div class="alert alert-success fade in">
+  		<button type="button" class="close" data-dismiss="alert" aria-hidden="true">×</button>
+  		{{ Session::get('message') }}
+		</div>
+	@endif
+@endif
 
 @if ($Precios->count())
 	<table class="table table-striped table-bordered">
@@ -14,6 +44,7 @@
 				<th>Preguntas</th>
 				<th>Panelistas</th>
 				<th>Precio</th>
+				<th>Estado</th>
 			</tr>
 		</thead>
 
@@ -24,18 +55,28 @@
 					<td>{{{ $Precio->preguntas }}}</td>
 					<td>{{{ $Precio->panelistas }}}</td>
 					<td>{{{ $Precio->precio }}}</td>
-                    <td>{{ link_to_route('Precios.edit', 'Edit', array($Precio->id), array('class' => 'btn btn-info')) }}</td>
+					<td>{{{ $Precio->activo == 1 ? 'Activo' : 'Inactivo' }}}</td>
+                    <td>{{ link_to_route('Precios.edit', 'Editar', array($Precio->id), array('class' => 'btn btn-info')) }}</td>
                     <td>
-                        {{ Form::open(array('method' => 'DELETE', 'route' => array('Precios.destroy', $Precio->id))) }}
-                            {{ Form::submit('Delete', array('class' => 'btn btn-danger')) }}
-                        {{ Form::close() }}
+                    	@if ($Precio->activo == 1)
+	                        {{ Form::open(array('method' => 'DELETE', 'route' => array('Precios.destroy', $Precio->id))) }}
+	                            {{ Form::submit('Desactivar', array('class' => 'btn btn-danger')) }}
+	                        {{ Form::close() }}
+	                    @else
+	                    	{{ Form::open(array('method' => 'POST', 'route' => array('Precios.activar', $Precio->id))) }}
+	                            {{ Form::submit('Activar', array('class' => 'btn btn-success')) }}
+	                        {{ Form::close() }}
+	                    @endif
                     </td>
 				</tr>
 			@endforeach
 		</tbody>
 	</table>
+	<div style="margin-left:-8%">{{$Precios->links()}}</div>
 @else
-	There are no Precios
+	<div class="alert alert-danger">
+	  <strong>Oh no!</strong> No hay Precios Disponibles
+	</div>
 @endif
 
 @stop
