+++ app/views/Sexos/index.blade.php
@@ -2,9 +2,39 @@
 
 @section('main')
 
-<h1>All Sexos</h1>
+<div class="page-header clearfix">
+    <h2 class="pull-left"><span class="glyphicon glyphicon-cog"></span> Sexos </h2>
+    <div class="pull-right">
+        <a href="{{{ URL::route('Configuracion') }}}" class="btn btn-sm btn-success"><span class="glyphicon glyphicon-arrow-left"></span> Regresar</a>
+    </div>
+</div>
 
-<p>{{ link_to_route('Sexos.create', 'Add new Sexo') }}</p>
+<div class="btn-agregar">
+	<a type="button" href="{{ URL::route('Sexos.create') }}" class="btn btn-primary">
+	  <span class="glyphicon glyphicon-file"></span> Nuevo Sexo
+	</a>
+</div>
+
+@if ($errors->any())
+  <div class="alert alert-danger fade in">
+    <button type="button" class="close" data-dismiss="alert" aria-hidden="true">×</button>
+    @if($errors->count() > 1)
+      <h4>Oh no! Se encontraron errores!</h4>
+    @else
+      <h4>Oh no! Se encontró un error!</h4>
+    @endif
+    <ul>
+      {{ implode('', $errors->all('<li class="error">:message</li>')) }}
+    </ul>  
+  </div>
+@else
+	@if (Session::has('message'))
+		<div class="alert alert-success fade in">
+  		<button type="button" class="close" data-dismiss="alert" aria-hidden="true">×</button>
+  		{{ Session::get('message') }}
+		</div>
+	@endif
+@endif
 
 @if ($Sexos->count())
 	<table class="table table-striped table-bordered">
@@ -12,7 +42,7 @@
 			<tr>
 				<th>Id</th>
 				<th>Nombre</th>
-				<th>Descripcion</th>
+				<th>Descripción</th>
 				<th>Activo</th>
 			</tr>
 		</thead>
@@ -23,19 +53,28 @@
 					<td>{{{ $Sexo->id }}}</td>
 					<td>{{{ $Sexo->nombre }}}</td>
 					<td>{{{ $Sexo->descripcion }}}</td>
-					<td>{{{ $Sexo->activo }}}</td>
-                    <td>{{ link_to_route('Sexos.edit', 'Edit', array($Sexo->id), array('class' => 'btn btn-info')) }}</td>
+					<td>{{{ $Sexo->activo == 1 ? 'Activo' : 'Inactivo' }}}</td>
+                    <td>{{ link_to_route('Sexos.edit', 'Editar', array($Sexo->id), array('class' => 'btn btn-info')) }}</td>
                     <td>
-                        {{ Form::open(array('method' => 'DELETE', 'route' => array('Sexos.destroy', $Sexo->id))) }}
-                            {{ Form::submit('Delete', array('class' => 'btn btn-danger')) }}
-                        {{ Form::close() }}
+                    	@if ($Sexo->activo == 1)
+	                        {{ Form::open(array('method' => 'DELETE', 'route' => array('Sexos.destroy', $Sexo->id))) }}
+	                            {{ Form::submit('Desactivar', array('class' => 'btn btn-danger')) }}
+	                        {{ Form::close() }}
+	                    @else
+	                    	{{ Form::open(array('method' => 'POST', 'route' => array('Sexos.activar', $Sexo->id))) }}
+	                            {{ Form::submit('Activar', array('class' => 'btn btn-success')) }}
+	                        {{ Form::close() }}
+	                    @endif
                     </td>
 				</tr>
 			@endforeach
 		</tbody>
 	</table>
+	<div style="margin-left:-8%">{{$Sexos->links()}}</div>
 @else
-	There are no Sexos
+	<div class="alert alert-danger">
+	  <strong>Oh no!</strong> No hay Sexos Disponibles
+	</div>
 @endif
 
 @stop
